#39 :: bugfix -- package GPIO renamed to gpio to follow naming conventions
